refactor(neo4j-storage): extract search mixin, finalize composition (issue #50 sub 3/3)
Issue #50 (EPIC-08-ST-01), Sub-Slice 3/3 — closes the issue, EPIC-08
and v0.9.0.

- New: backend/app/storage/neo4j_search.py (57 LOC)
  - Neo4jSearchMixin holds the single search(graph_id, query, limit,
    scope) hybrid-search entry point. Vector/keyword logic itself
    stays in storage.search_service via self._search.
- Changed: backend/app/storage/neo4j_storage.py 228 -> 195 LOC (-33)
  - Final composition:
    class Neo4jStorage(Neo4jReadMixin, Neo4jWriteMixin,
                       Neo4jSearchMixin, GraphStorage)
  - search method removed. typing.Any/typing.Dict imports drop.
  - Remaining in the file: __init__/close, ontology-service late-bind,
    _verify_connectivity, _ensure_schema, health-status properties
    (is_connected, last_error, last_success_ts) for /api/status,
    _call_with_retry with health-state tracking, and the
    staticmethod re-exports of _node_to_dict/_edge_to_dict.

Acceptance criteria fully met:
- [x] "Schreiblogik, Leselogik und Suche liegen nicht mehr in einer
  Datei" — three separate modules + mixins
- [x] "Mappings sind zentral wiederverwendbar" — neo4j_mappings.py

LOC for issue #50 in total: 1127 -> 195 (-932, -82.7%, four fifths
gone). Distributed: 195 + 381 + 513 + 57 + 118 = 1264 LOC across 5
files. The +137 increase over the inline version comes from module
docstrings, typed mixin signatures and re-export comments — explicit
API documentation, not duplicated logic.

EPIC-08 closed (3/3: #50, #51, #52).
v0.9.0 "Domain Cleanup" milestone closed (12/12).

LOC reductions on the three v0.9.0 hot-spots:
- simulation_manager.py: 789 -> 403 (-49%)
- report_agent.py:       3184 -> 2179 (-31.6%)
- neo4j_storage.py:      1127 -> 195 (-82.7%)

Test counter: 531 -> 671 (+140 in v0.9.0 path A).

npm run check: 5/5 green (671 backend, 40 frontend, build ok).

Closes #50
